requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
PyYAML==6.0.1
python-socketio==5.10.0
picamera2==0.3.12
RPi.GPIO==0.7.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
eventlet==0.33.3

